+++ lib/zenlish/lex/empty_lexicon_factory.rb
@@ -18,8 +18,8 @@ module Zenlish
         aLexicon.add_terminal(WClasses::Adjective.new.freeze)
         aLexicon.add_terminal(WClasses::Adverb.new.freeze)
         aLexicon.add_terminal(WClasses::AdverbNot.new.freeze)
-        aLexicon.add_terminal(WClasses::AdverbThere.new.freeze)        
-        # aLexicon.add_terminal(WClasses::AuxiliaryBe.new.freeze)
+        aLexicon.add_terminal(WClasses::AdverbThere.new.freeze)
+        aLexicon.add_terminal(WClasses::AuxiliaryBe.new.freeze)
         aLexicon.add_terminal(WClasses::AuxiliaryDo.new.freeze)        
         aLexicon.add_terminal(WClasses::Cardinal.new.freeze)
         aLexicon.add_terminal(WClasses::CommonNoun.new.freeze)
@@ -30,14 +30,19 @@ module Zenlish
         aLexicon.add_terminal(WClasses::DemonstrativePronoun.new.freeze)        
         aLexicon.add_terminal(WClasses::IndefinitePronoun.new.freeze)
         aLexicon.add_terminal(WClasses::IrregularVerb.new.freeze)
-        aLexicon.add_terminal(WClasses::IrregularVerbBe.new.freeze)        
+        aLexicon.add_terminal(WClasses::IrregularVerbBe.new.freeze)   
+        aLexicon.add_terminal(WClasses::LinkingAdverb.new.freeze)        
         aLexicon.add_terminal(WClasses::Preposition.new.freeze)
-        aLexicon.add_terminal(WClasses::PrepositionThan.new.freeze)        
+        aLexicon.add_terminal(WClasses::PrepositionOf.new.freeze)        
+        aLexicon.add_terminal(WClasses::PrepositionThan.new.freeze) 
+        aLexicon.add_terminal(WClasses::RegularVerb.new.freeze)         
         aLexicon.add_terminal(WClasses::ProperNoun.new.freeze)
         aLexicon.add_terminal(WClasses::IndefiniteQuantifier.new.freeze)
+        aLexicon.add_terminal(WClasses::SubordinatingConjunction.new.freeze)        
       end
 
       def add_punctuation(aLexicon)
+        aLexicon.add_terminal(Rley::Syntax::Terminal.new('Comma'))
         aLexicon.add_terminal(Rley::Syntax::Terminal.new('Period'))
       end
     end # module