+++ lib/zenlish/wclasses/linking_adverb.rb
@@ -1,0 +1,9 @@
+require_relative 'adverb'
+
+module Zenlish
+  module WClasses
+    # TODO: e.g. then
+    class LinkingAdverb < Adverb
+    end # class
+  end # module
+end # module