+++ lib/zenlish/wclasses/regular_verb.rb
@@ -1,0 +1,9 @@
+require_relative 'lexical_verb'
+
+module Zenlish
+  module WClasses
+    # TODO: document 'touch'
+    class RegularVerb < LexicalVerb
+    end # class
+  end # module
+end # module