+++ lib/zenlish/wclasses/subordinating_conjunction.rb
@@ -1,0 +1,9 @@
+require_relative 'conjunction'
+
+module Zenlish
+  module WClasses
+    # TODO: if, 
+    class SubordinatingConjunction < WordClass
+    end # class
+  end # module
+end # module